bugfix: report error if omfile cannot open file
The current state is incomplete. Most importantly, it affects all types
of files (like queues), without us checking the side-effects of this.

Further implementation is currently on hold due to this question from
the mailing list:

>> I would have expected rsyslog to show errors in it's logs and/or problems
>> in impstats when maxopenfiles is hit and it can't open a file for output.
>
> ACK,  that strongly smells like a bug.

I have looked into it, and I now think I know why no error message was
generated: this can lead to an error message loop. This is the case if
the error message is written to a file that itself had the error. So
the next write will also generate one and then rsyslog is busy
reporting the error that happened during error reporiting ... and so
on. There is the default rate-limiter which places some upper bound on
the iterations, but this still can get very ugly.

Note that tracking the "already reported" state for each file is more
or less out of question -- that would mean we would need to maintain a
list for all files that ever have been tried. Or is that acceptable?
In any case, this means we need to add a lot of code for this error
tracking.