Get the new test to run without hanging
The new test now runs without hanging, but it still fails because while
the .conf file parser is case-insensitive, it does not change the case of
the references parsed. So the variables are used as is.  The only place
(that I could find) which changes the case of variables is in templates,
where variable references are forced lower case.