imudp: add one retry on send failure due to EAGAIN
This can happen when the UDP sendbuffer is full and might actually happen
on busy systems (with busy networks). So far, this always led to action
suspension, which is a bit drastic. This patch now does a short wait and
one retry. This is along the same line of what is done for TCP.

Right now, there are no new config parameters. Should need arise to permit
configuration, we can always add them.

While the timing somewhat changes, we do not expect this is notably and,
more importantly, we do not think that this can have any effect on
current deployments (except that they may run smoother).