core: consistent handling of oversize input messages
In the community we frequently discuss handling of oversize messages.
David Lang rightfully suggested to create a central capability inside
rsyslog core to handle them.

We need to make a distinction between input and output messages. Also,
input messages frequently need to have some size restrictions done at
a lower layer (e.g. protocol layer) for security reasons. Nevertheless,
we should have a central capability

* for cases where it need not be handled at a lower level
* as a safeguard when a module invalidly emits it (imfile is an example,
  see https://github.com/rsyslog/rsyslog/pull/2632 for a try to fix it
  on the module level - we will replace that with the new capability
  described here).

The central capability works on message submission, and so cannot be
circumvented. It needs to have at least these capabilities:

* truncate message
* split message
  this is of questionable use, but also often requested. In that mode,
  the oversize message content is split into multiple messages. Usually,
  this ends up with message segments where all but the first is lost
  anyhow as the regular filter rules do not match the other fragments.
  As it is requested, we still implemented it.
* report message to a special "oversize message log file" (not via the
  regular engine, as that would obviously cause another oversize message)

It is questionable if we should also permit a mode

* accept

where the oversize message is accepted as-is. This will be part of another
commit if we decide to do so.

This commit, as the title says, handles oversize INPUT messages.

see also https://github.com/rsyslog/rsyslog/issues/2190
closes https://github.com/rsyslog/rsyslog/issues/2681
closes https://github.com/rsyslog/rsyslog/issues/498

Note: this commit adds global parameters:
 * "oversizemsg.errorfile",
   is used to specify the location of the oversize message log file.
 * add global parameter "oversizemsg.input.mode"
   is used to specify the mode with which oversized messages will
   be handled.