WiP: imtcp bugfix: memory leak on shutdown (cosmetic)
when rsyslog shuts down and imtcp has still data it is unable to
process and imtcp is using epoll mode, the session object is not
destructed before shutdown.

This is a cosmetic leak, but brings up memory analyzer errors in CI.

WE NEED TO CHECK IF IT IS WORTH FIXING IT, as some runtime overhead
is involved.