bugfix: ZSTD_e_end requries different loop condition check
The predicate check for termination of the write loop must be different
when ZSTD_e_end is used. This was done incorrectly in the main writer
loop, resulting in "Operation not authorized at current processing stage"
occasional error messages.

Note: this commit's code fixes the problem but uses ZSTD_e_flush instead
(where the problem does not exist). This is in order to aid some performance
testing. The decision about final modes is still to be made.

INTERIM COMMIT - NEEDS FOLLOW UP!