various ways to handle synchronization accept/regular workers
this looks like a dead end and hard to solve. In conclusion, I
think there is no value in splitting of the accept() part of
the server in our rsyslog case (we are not a web server!).

Nevertheless, I would like to preserve this state of code
changes.

Please note that I can carry over much of the work into the
approach with combined event loop (again), we learned a lot
by exploring the dedicated accept worker pool.